requirements.txt
# requirements.txt
torch>=1.9.0
torchvision>=0.10.0
mmcv-full>=1.3.13,<1.8.0
mmsegmentation>=0.28.0
opencv-python>=4.5.0
matplotlib>=3.3.0
seaborn>=0.11.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
Pillow>=8.3.0
tqdm>=4.62.0
jupyter>=1.0.0
ipykernel>=6.0.0
plotly>=5.0.0
openpyxl>=3.0.7
xlsxwriter>=3.0.0
psutil>=5.8.0
GPUtil>=1.4.0



